your view of all the short links

+++ src/lib/db/linkActions.ts
@@ -1,24 +1,47 @@
 "use server";
 
 import { db } from "@/db/drizzle";
-import { eq, sql } from "drizzle-orm";
-import { link } from "@/db/schema";
+import { eq, sql, and } from "drizzle-orm";
+import { link, user } from "@/db/schema";
 import { randomUUID } from "crypto";
 
-export const generateLink = async (url: string, shortUrl: string, collectStats: boolean) => {
+export const generateLink = async (
+  url: string,
+  shortUrl: string,
+  collectStats: boolean,
+  author: string
+) => {
   const newLink = await db.insert(link).values({
     id: randomUUID(),
     url,
     shortUrl,
-    collectStats
+    collectStats,
+    author,
   }).returning();
 
-  return newLink[0];
+  const updatedUser = await db.update(user)
+    .set({
+      shortLinks: sql`array_append(${user.shortLinks}, ${shortUrl})`,
+    })
+    .where(eq(user.id, author))
+    .returning();
+
+  return { newLink: newLink[0], updatedUser: updatedUser[0] };
 }
 
 export const getLinkByShortUrl = async (shortUrl: string) => {
   const foundLink = await db.select().from(link).where(eq(link.shortUrl, shortUrl));
   return foundLink[0];
+}
+
+export const getLinkStats = async (shortUrl: string, userId: string) => {
+  const foundLink = await db.select().from(link).where(and(eq(link.shortUrl, shortUrl), eq(link.author, userId)));
+  return foundLink[0];
+}
+
+export const getLinksByUserId = async (userId: string) => {
+  const links = await db.select().from(link).where(eq(link.author, userId));
+  return links;
 }
 
 export const incrementAccessCount = async (shortUrl: string) => {
@@ -33,9 +56,9 @@ export const incrementAccessCount = async (shortUrl: string) => {
   return updatedLink[0];
 }
 
-export const deleteLinkByShortUrl = async (shortUrl: string) => {
+export const deleteLinkByShortUrl = async (shortUrl: string, userId: string) => {
   const deletedLink = await db.delete(link)
-    .where(eq(link.shortUrl, shortUrl))
+    .where(and(eq(link.shortUrl, shortUrl), eq(link.author, userId)))
     .returning();
 
   return deletedLink[0];
